Simplify TM grammar, add it to VSCode extension for theming where semantic tokens are unavailable (#2276)
* Simplify textmate grammar, configure VSCode extension for hover highlighting

* Fix bad comment handling, add baselines for it

* Update lockfile

* How could I forget you, existing?
