+++ src/highlightjs/jest.config.ts
@@ -10,7 +10,7 @@ module.exports = {
     '^.+\\.(ts|tsx)$': 'ts-jest'
   },
   testMatch: [
-    '**/test/**/*.test.(ts|js)'
+    '**/test/**/*.test.(ts)'
   ],
   testEnvironment: 'node',
 };
