+++ src/highlightjs/src/bicep.ts
@@ -14,7 +14,7 @@ const identifierContinue = "[_a-zA-Z0-9]";
 const identifier = bounded(`${identifierStart}${identifierContinue}*`);
 
 // whitespace. ideally we'd tokenize in-line block comments, but that's a lot of work. For now, ignore them.
-const ws = `(?:\\s|/\\*.*\\*/)*`;
+const ws = `(?:[ \\t\\r\\n]|\\/\\*(?:\\*(?!\\/)|[^*])*\\*\\/)*`;
 
 const KEYWORDS = {
   $pattern: '[A-Za-z$_][0-9A-Za-z$_]*',
@@ -28,6 +28,7 @@ const KEYWORDS = {
     'for',
     'in',
     'if',
+    'existing',
   ],
   literal: [
     "true",
