+++ src/highlightjs/test/baselines/basic.html
@@ -62,6 +62,10 @@
   }
 }]
 
+<span class="hljs-variable"><span class="hljs-keyword">resource</span></span> <span class="hljs-variable">virtualNetwork</span> <span class="hljs-string">&#x27;Microsoft.Network/virtualNetworks@2020-08-01&#x27;</span> <span class="hljs-variable"><span class="hljs-keyword">existing</span></span> = {
+  <span class="hljs-property">name</span>: <span class="hljs-string">&#x27;myVnet&#x27;</span>
+}
+
 <span class="hljs-variable"><span class="hljs-keyword">var</span></span> <span class="hljs-variable">multi</span> = <span class="hljs-string">&#x27;&#x27;&#x27;&#x27;&#x27;&#x27;</span>
 <span class="hljs-variable"><span class="hljs-keyword">var</span></span> <span class="hljs-variable">multi2</span> = <span class="hljs-string">&#x27;&#x27;&#x27;
       hello!
