+++ src/highlightjs/test/baselines/comments.html
@@ -1,0 +1,23 @@
+
+<html>
+  <head>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.7.2/styles/default.min.css">
+  </head>
+  <body>
+    <pre class="hljs">
+<span class="hljs-variable"><span class="hljs-keyword">resource</span></span> <span class="hljs-variable">test</span> <span class="hljs-string">&#x27;Microsoft.AAD/domainServices@2021-03-01&#x27;</span> = {
+  <span class="hljs-property">name</span>: <span class="hljs-string">&#x27;asdfsdf&#x27;</span>
+  <span class="hljs-comment">// this is a comment</span>
+  <span class="hljs-property">properties</span>: {<span class="hljs-comment">/*comment*/</span>
+    <span class="hljs-property">domainConfigurationType</span><span class="hljs-comment">/*comment*/</span>:<span class="hljs-comment">/*comment*/</span><span class="hljs-string">&#x27;as//notacomment!d/* also not a comment */fsdf&#x27;</span><span class="hljs-comment">// test!/*</span>
+    <span class="hljs-comment">/* multi
+    line
+    comment */</span> <span class="hljs-property">domainName</span>: <span class="hljs-comment">/*
+    asdf*/</span><span class="hljs-string">&#x27;test&#x27;</span>
+    <span class="hljs-comment">// comment</span>
+  }
+}
+
+    </pre>
+  </body>
+</html>
