+++ src/textmate/package.json
@@ -3,6 +3,7 @@
   "private": true,
   "devDependencies": {
     "@azure-tools/tmlanguage-generator": "^0.1.3",
+    "@types/html-escaper": "^3.0.0",
     "@types/jest": "^26.0.22",
     "@types/node": "^14.14.39",
     "@types/plist": "^3.0.2",
@@ -11,6 +12,7 @@
     "eslint": "^7.24.0",
     "eslint-plugin-header": "^3.1.1",
     "eslint-plugin-jest": "^24.3.5",
+    "html-escaper": "^2.0.2",
     "jest": "^26.6.3",
     "onigasm": "^2.2.5",
     "plist": "^3.0.2",
