--- src/textmate/test/baselines/bad_inlinecomments.bicep
@@ -1,3 +1,0 @@
-var test = {
-  /* this doesn't work :( */
-}
