--- src/textmate/test/baselines/bad_partial.bicep
@@ -1,3 +1,0 @@
-resource partialDeclaration
-
-resource next
