--- src/textmate/test/baselines/bad_partial.tokens
@@ -1,4 +1,0 @@
-  resource partialDeclaration
-  
-  resource next
-
