+++ src/textmate/test/baselines/comments.html
@@ -1,0 +1,24 @@
+
+<html>
+  <head>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/10.7.2/styles/default.min.css">
+  </head>
+  <body>
+    <pre class="hljs">
+<span class="hljs-keyword">resource</span> <span class="hljs-variable">test</span> <span class="hljs-string">&#39;</span><span class="hljs-string">Microsoft.AAD/domainServices@2021-03-01</span><span class="hljs-string">&#39;</span> = {
+  <span class="hljs-property">name</span>: <span class="hljs-string">&#39;</span><span class="hljs-string">asdfsdf</span><span class="hljs-string">&#39;</span>
+  <span class="hljs-comment">// this is a comment</span>
+  <span class="hljs-property">properties</span>: {<span class="hljs-comment">/*</span><span class="hljs-comment">comment</span><span class="hljs-comment">*/</span>
+    <span class="hljs-property">domainConfigurationType</span><span class="hljs-comment">/*</span><span class="hljs-comment">comment</span><span class="hljs-comment">*/</span>:<span class="hljs-comment">/*</span><span class="hljs-comment">comment</span><span class="hljs-comment">*/</span><span class="hljs-string">&#39;</span><span class="hljs-string">as//notacomment!d/* also not a comment */fsdf</span><span class="hljs-string">&#39;</span><span class="hljs-comment">// test!/*</span>
+    <span class="hljs-comment">/*</span><span class="hljs-comment"> multi</span>
+<span class="hljs-comment">    line</span>
+<span class="hljs-comment">    comment </span><span class="hljs-comment">*/</span> <span class="hljs-property">domainName</span>: <span class="hljs-comment">/*</span>
+<span class="hljs-comment">    asdf</span><span class="hljs-comment">*/</span><span class="hljs-string">&#39;</span><span class="hljs-string">test</span><span class="hljs-string">&#39;</span>
+    <span class="hljs-comment">// comment</span>
+  }
+}
+
+
+    </pre>
+  </body>
+</html>
