+++ src/vscode-bicep/package.json
@@ -79,11 +79,29 @@
         ],
         "configuration": "./language-configuration.json"
       }
+    ],
+    "grammars": [
+      {
+        "scopeName": "markdown.bicep.codeblock",
+        "path": "./codeblock.json",
+        "injectTo": [
+          "text.html.markdown"
+        ],
+        "embeddedLanguages": {
+          "meta.embedded.block.bicep": "bicep"
+        }
+      },
+      {
+        "language": "bicep",
+        "scopeName": "source.bicep",
+        "path": "./bicep.tmlanguage"
+      }
     ]
   },
   "scripts": {
     "vscode:prepublish": "webpack --mode production",
-    "compile": "tsc -p ./",
+    "compile": "tsc -p ./ && npm run copy:tmlanguage",
+    "copy:tmlanguage": "ncp ../textmate/bicep.tmlanguage ./bicep.tmlanguage",
     "watch": "tsc -watch -p ./",
     "lint": "eslint src --ext ts",
     "lint:fix": "eslint src --ext ts --fix",
@@ -93,6 +111,7 @@
     "package": "npm run clean && nbgv-setversion && vsce package --githubBranch main --out ./vscode-bicep.vsix && nbgv-setversion --reset"
   },
   "devDependencies": {
+    "@types/copy-webpack-plugin": "^6.4.1",
     "@types/jest": "^26.0.22",
     "@types/mocha": "^8.2.2",
     "@types/node": "^14.14.39",
@@ -102,12 +121,14 @@
     "@types/webpack": "^5.28.0",
     "@typescript-eslint/eslint-plugin": "^4.22.0",
     "@typescript-eslint/parser": "^4.22.0",
+    "copy-webpack-plugin": "^8.1.1",
     "eslint": "^7.24.0",
     "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-header": "^3.1.1",
     "eslint-plugin-jest": "^24.3.5",
     "eslint-plugin-prettier": "^3.4.0",
     "jest": "^26.6.2",
+    "ncp": "^2.0.0",
     "nerdbank-gitversioning": "^3.4.194",
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
